Work around Safari flexbox bugs
Safari compresses the nav-menu if "flex-shrink-0" is not set in Bootstrap. It also does not expand the 100% height of the flexbox child element that is used to evenly space out the author information and the post content when short posts (link, video, micro) are displayed.

Chrome and Firefox rendered everything correctly.