Remove pagination and jumbotron
Remove the pagination as it does not improve the front page. Last 12
entries are displayed, no matter if status update or fully fledged
post. Maybe in the future i will create a two or three column design
that lists each type separately. But not for now, people interested
can click on "Posts" and "Updates".

Remove the jumbotron as it makes the text less readable, especially
for visually impaired people. I decided to increase the font-weight
for the lead text for the same reason.